% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb2long.R
\name{ukb_fieldarray2long}
\alias{ukb_fieldarray2long}
\title{convert into 'superlong' format with separate row for
each array.}
\usage{
ukb_fieldarray2long(ukb)
}
\description{
convert into 'superlong' format with separate row for
each array.
}

