% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb2long.R
\name{ukb_fields2long_single}
\alias{ukb_fields2long_single}
\title{turn a single f.eid, f.1234.5.6 pair into long format}
\usage{
ukb_fields2long_single(ukb)
}
\description{
turn a single f.eid, f.1234.5.6 pair into long format
}

