% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb2long.R
\name{ukb_fields2long_type}
\alias{ukb_fields2long_type}
\title{Given a data.frame of UKB data with an f.eid column and
other columns of the same type (numeric, factor, character)}
\usage{
ukb_fields2long_type(ukb)
}
\description{
Given a data.frame of UKB data with an f.eid column and
other columns of the same type (numeric, factor, character)
}

